Add Modules For Control
Add Modules for each function to control how the robot operates. Each
module is responsible for its own function and is mapped to its own
button.  Drive Modules determine what type of drive is being used and
which motors are operated as well as which axes control the motor
operation.