Swerve Tweaks - Stuttering
Moves zeroing for deceleration to motor speed rather than delta angle,
as delta angle seroing was skewing the decelaration calculations
Added value inversion to controller to allow for inverting thumbsticks.
Inverted rotation motor to account for encoder spinning the opposite
direction of the motor when using the analog absolute encoder.