Revert back to division and elevator fixes
With Radians Per Second values below 50, using multiplication can
'appear' to be correct because it results in a value under 1.  After
attempting values over 50 (594), it became apparent that the math was
incorrect.  Multiplying results in values that are much too high for RPS
over 50 and much too low for RPS under 50.  Division is correct for this
calculation.

Updated Encoder functions for Elevator, as changes to swerve affected
the rev encoder class that elevator depended on for counting rotations.